Final - AI, Collision fix, Smash, Serve
AI - basic ai to follow the ball based on its Y position.
Collision fix - Paddles and ball moving out of the canvas.
Smash - whenever a paddle hits the ball on the very end, the speed of
the ball multiplies by 1.5x.
Serve - whenever a player or cpu wins, reset the ball and serve at a
random X and Y